Stop trimming at a certain point
Trimming isomorphic moves at each node in the entire tree is quite
a bit slower than not even trimming it all. Instead, once a node
is found to have no isomorphic children, then stop trimming for all
of its descendants down the tree. This heuristic still allows trimming
at the top of the tree where trimming is most effective, and then phases
it out as the depth increases. This leads to about a 2x speedup over not
trimming at all.